Create SamplePhotoViewController
VC shows a video stream of the camera input and has a button so that the user can take a snapshot and save that image in the local storage.

This feature will be used to take lots of photos of the furniture, apply object detection, crop out images containing the objects and send those images to a python script which will train a model.