Split Test and Training data
This split makes it so that no augmented images from the train set
are tested on. This is not good since it gives a false impression
that the model is really good. The model has already seen the
original images during training and the modified are thus easier
to recognize.