Remove the object tracker from the application
The object tracker together with the connecting arrow
created a positive feedback loop which made the app
unstable.

Instead of tracking object around, an assumption is made
that the furniture part objects will always stay in the same
world position when holding the camera.
Thus, the objects are in no need of being track but rather the
AR Session will keep track of the position.